slack/pyproject.toml
[project]
name = "thufir-slack"
version = "0.1.0"
description = "Slack bot for the Thufir data-retrieval agent"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]

